Add new build/release action * Add vendored-openssl feature flag
This allows us to trigger a build which
includes a static dependency to openssl by compiling it from source in a
build.rs script. It is useful for Linux, which has wildly different config
parameters for different architectures.

* Add openssl-sys as direct dependency

This allows us to pass the vendored-openssl
feature flag to this dependency and compile openssl from source